Containerize the runtime events in a ring buffer.
To enable someone to use this tool to observe event streams over
sliding windows of sample recency as well as prevent unbound growth of
observations, the list of observations has been wrapped in a ring
buffer of flag-specifiable size.  This has been a tremendous help
in visualizing events from particularly garbage collector-active
servers.